app: two-factor-auth: never compare boxed Integers with operators
There is no operator overloading in Java. This was doing reference
equality checks, which is guaranteed to never be true.

+++ src/app/twofactorauth/QrCodeUtil.java
@@ -17,7 +17,7 @@ public class QrCodeUtil {
         InputStream inputStream = connection.getInputStream();
         byte[] qrImageData = new byte[contentLength];
         try {
-            while(readBytes != contentLength) {
+            while(readBytes.intValue() != contentLength.intValue()) {
                 Integer rb = inputStream.read(qrImageData, readBytes, contentLength - readBytes);
                 if(rb < 0) {
                     break;
